media-source/MediaSource-addSourceBuffer_null.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>MediaSource#addSourceBuffer(null) and MediaSource#addSourceBuffer("")</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>

<div id="log"></div>

<script>
var FILE = '/media/white.webm';
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vorbis,vp8"';
async_test(function(t) {
    var mediaSource = new MediaSource();

    mediaSource.addEventListener('sourceopen', function(e){
        var sourceBuffer = null;
        t.step(function() {
            assert_throws({name : 'InvalidAccessError'},
                          function() {
                              sourceBuffer = mediaSource.addSourceBuffer(null);
                          },
                          'Set null');
        });
        t.step(function() {
            assert_throws({name : 'InvalidAccessError'},
                          function() {
                              sourceBuffer = mediaSource.addSourceBuffer('');
                          },
                          'Set a empty string');
        });
        t.done();
        mediaSource.endOfStream();
    }, false);
    video.src = window.URL.createObjectURL(mediaSource);
}, 'MediaSource#addSourceBuffer : If type is null or an empty string then throw an INVALID_ACCESS_ERR exception and abort these steps.');

</script>
</body>
</html>
